requirements.txt
b3j0f.conf==0.3.19
b3j0f.utils==1.4.3
link.crdt==1.0
link.feature==2.1
link.middleware==1.5
link.kvstore==1.0
link.fulltext==0.8
six==1.10.0
riak==2.5.5
